[BugFix] Ensure src is only set if defined
[fixes #27]

Issue is caused by some browser specific implementation details around
`new Audio`.  In Chrome when you create a new `Audio` instance passing
in undefined as `src` it ostensibly ignores it.  Under the preceding
circumstance when you request `(new Audio(undefined)).src)` you will
receive `""` in Chrome, but `http://localhost:4200/undefined` in Safari.

Later in the `player` service `select` method we test for the presence
of `src` to determine whether or not we should set it.  Safari's
behavior caused that test to produce incorrect results.

Thanks for /r/emberjs redditor Limonero for pointing out this error in
Safari!